pyproject.toml
[tool.poetry]
name = "qa-wikidata"
version = "0.1.0"
description = ""
authors = ["Adrian Moses <adrian@adrianmoses.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
datasets = "^2.16.1"
openai = "^1.10.0"
tqdm = "^4.66.1"
pinecone-client = "^3.0.2"
pandas = "^2.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
